style - consistent nameing and style for gpu backends
