+++ backends/cuda-gen/ceed-cuda-gen-operator-build.cpp
@@ -23,7 +23,7 @@
 #include <sstream>
 #include "ceed-cuda-gen.h"
 #include "../cuda/ceed-cuda-compile.h"
-#include "../cuda-ref/ceed-cuda.h"
+#include "../cuda-ref/ceed-cuda-ref.h"
 #include "../cuda-shared/ceed-cuda-shared.h"
 
 static const char *atomicAdd = QUOTE(
@@ -803,8 +803,8 @@ extern "C" int CeedCudaGenOperatorBuild(CeedOperator op) {
   struct cudaDeviceProp prop;
   Ceed_Cuda *ceed_data;
   ierr = CeedGetData(ceed, &ceed_data); CeedChkBackend(ierr); CeedChkBackend(ierr);
-  ierr = cudaGetDeviceProperties(&prop, ceed_data->deviceId); CeedChkBackend(ierr);
-  if ((prop.major<6) && (CEED_SCALAR_TYPE != CEED_SCALAR_FP32)){
+  ierr = cudaGetDeviceProperties(&prop, ceed_data->device_id); CeedChkBackend(ierr);
+  if ((prop.major < 6) && (CEED_SCALAR_TYPE != CEED_SCALAR_FP32)){
     code << atomicAdd;
   }
