+++ backends/cuda-ref/ceed-cuda-ref-qfunctioncontext.c
@@ -18,7 +18,7 @@
 #include <ceed/backend.h>
 #include <cuda_runtime.h>
 #include <string.h>
-#include "ceed-cuda.h"
+#include "ceed-cuda-ref.h"
 
 //------------------------------------------------------------------------------
 // * Bytes used
