+++ backends/cuda-ref/ceed-cuda-ref.h
@@ -14,14 +14,13 @@
 // software, applications, hardware, advanced system engineering and early
 // testbed platforms, in support of the nation's exascale computing imperative.
 
-#ifndef _ceed_hip_h
-#define _ceed_hip_h
+#ifndef _ceed_cuda_h
+#define _ceed_cuda_h
 
 #include <ceed/ceed.h>
 #include <ceed/backend.h>
-#include <hip/hip_runtime.h>
-#include <hipblas.h>
-#include "../hip/ceed-hip-common.h"
+#include <cuda.h>
+#include "../cuda/ceed-cuda-common.h"
 
 typedef struct {
   CeedScalar *h_array;
@@ -30,14 +29,14 @@ typedef struct {
   CeedScalar *d_array;
   CeedScalar *d_array_borrowed;
   CeedScalar *d_array_owned;
-} CeedVector_Hip;
+} CeedVector_Cuda;
 
 typedef struct {
-  hipModule_t module;
-  hipFunction_t noTrStrided;
-  hipFunction_t noTrOffset;
-  hipFunction_t trStrided;
-  hipFunction_t trOffset;
+  CUmodule module;
+  CUfunction noTrStrided;
+  CUfunction noTrOffset;
+  CUfunction trStrided;
+  CUfunction trOffset;
   CeedInt nnodes;
   CeedInt *h_ind;
   CeedInt *h_ind_allocated;
@@ -46,23 +45,23 @@ typedef struct {
   CeedInt *d_toffsets;
   CeedInt *d_tindices;
   CeedInt *d_lvec_indices;
-} CeedElemRestriction_Hip;
+} CeedElemRestriction_Cuda;
 
 // We use a struct to avoid having to memCpy the array of pointers
 // __global__ copies by value the struct.
 typedef struct {
   const CeedScalar *inputs[CEED_FIELD_MAX];
   CeedScalar *outputs[CEED_FIELD_MAX];
-} Fields_Hip;
+} Fields_Cuda;
 
 typedef struct {
-  hipModule_t module;
+  CUmodule module;
   char *qFunctionName;
   char *qFunctionSource;
-  hipFunction_t qFunction;
-  Fields_Hip fields;
+  CUfunction qFunction;
+  Fields_Cuda fields;
   void *d_c;
-} CeedQFunction_Hip;
+} CeedQFunction_Cuda;
 
 typedef struct {
   void *h_data;
@@ -71,32 +70,32 @@ typedef struct {
   void *d_data;
   void *d_data_borrowed;
   void *d_data_owned;
-} CeedQFunctionContext_Hip;
+} CeedQFunctionContext_Cuda;
 
 typedef struct {
-  hipModule_t module;
-  hipFunction_t interp;
-  hipFunction_t grad;
-  hipFunction_t weight;
+  CUmodule module;
+  CUfunction interp;
+  CUfunction grad;
+  CUfunction weight;
   CeedScalar *d_interp1d;
   CeedScalar *d_grad1d;
   CeedScalar *d_qweight1d;
-} CeedBasis_Hip;
+} CeedBasis_Cuda;
 
 typedef struct {
-  hipModule_t module;
-  hipFunction_t interp;
-  hipFunction_t grad;
-  hipFunction_t weight;
+  CUmodule module;
+  CUfunction interp;
+  CUfunction grad;
+  CUfunction weight;
   CeedScalar *d_interp;
   CeedScalar *d_grad;
   CeedScalar *d_qweight;
-} CeedBasisNonTensor_Hip;
+} CeedBasisNonTensor_Cuda;
 
 typedef struct {
-  hipModule_t module;
-  hipFunction_t linearDiagonal;
-  hipFunction_t linearPointBlock;
+  CUmodule module;
+  CUfunction linearDiagonal;
+  CUfunction linearPointBlock;
   CeedBasis basisin, basisout;
   CeedElemRestriction diagrstr, pbdiagrstr;
   CeedVector elemdiag, pbelemdiag;
@@ -104,7 +103,7 @@ typedef struct {
   CeedEvalMode *h_emodein, *h_emodeout;
   CeedEvalMode *d_emodein, *d_emodeout;
   CeedScalar *d_identity, *d_interpin, *d_interpout, *d_gradin, *d_gradout;
-} CeedOperatorDiag_Hip;
+} CeedOperatorDiag_Cuda;
 
 typedef struct {
   CeedVector *evecs;   // E-vectors, inputs followed by outputs
@@ -114,27 +113,27 @@ typedef struct {
   CeedInt    numeout;
   CeedInt    qfnumactivein, qfnumactiveout;
   CeedVector *qfactivein;
-  CeedOperatorDiag_Hip *diag;
-} CeedOperator_Hip;
+  CeedOperatorDiag_Cuda *diag;
+} CeedOperator_Cuda;
 
-CEED_INTERN int CeedHipGetHipblasHandle(Ceed ceed, hipblasHandle_t *handle);
+CEED_INTERN int CeedCudaGetCublasHandle(Ceed ceed, cublasHandle_t *handle);
 
-CEED_INTERN int CeedVectorCreate_Hip(CeedInt n, CeedVector vec);
+CEED_INTERN int CeedVectorCreate_Cuda(CeedInt n, CeedVector vec);
 
-CEED_INTERN int CeedElemRestrictionCreate_Hip(CeedMemType mtype,
+CEED_INTERN int CeedElemRestrictionCreate_Cuda(CeedMemType mtype,
     CeedCopyMode cmode, const CeedInt *indices, CeedElemRestriction r);
 
-CEED_INTERN int CeedElemRestrictionCreateBlocked_Hip(const CeedMemType mtype,
+CEED_INTERN int CeedElemRestrictionCreateBlocked_Cuda(const CeedMemType mtype,
     const CeedCopyMode cmode, const CeedInt *indices,
     const CeedElemRestriction res);
 
-CEED_INTERN int CeedBasisApplyElems_Hip(CeedBasis basis, const CeedInt nelem,
-                                        CeedTransposeMode tmode, CeedEvalMode emode, const CeedVector u, CeedVector v);
+CEED_INTERN int CeedBasisApplyElems_Cuda(CeedBasis basis, const CeedInt nelem,
+    CeedTransposeMode tmode, CeedEvalMode emode, const CeedVector u, CeedVector v);
 
-CEED_INTERN int CeedQFunctionApplyElems_Hip(CeedQFunction qf, const CeedInt Q,
+CEED_INTERN int CeedQFunctionApplyElems_Cuda(CeedQFunction qf, const CeedInt Q,
     const CeedVector *const u, const CeedVector *v);
 
-CEED_INTERN int CeedBasisCreateTensorH1_Hip(CeedInt dim, CeedInt P1d,
+CEED_INTERN int CeedBasisCreateTensorH1_Cuda(CeedInt dim, CeedInt P1d,
     CeedInt Q1d,
     const CeedScalar *interp1d,
     const CeedScalar *grad1d,
@@ -142,16 +141,16 @@ CEED_INTERN int CeedBasisCreateTensorH1_Hip(CeedInt dim, CeedInt P1d,
     const CeedScalar *qweight1d,
     CeedBasis basis);
 
-CEED_INTERN int CeedBasisCreateH1_Hip(CeedElemTopology, CeedInt, CeedInt,
-                                      CeedInt, const CeedScalar *,
-                                      const CeedScalar *, const CeedScalar *,
-                                      const CeedScalar *, CeedBasis);
+CEED_INTERN int CeedBasisCreateH1_Cuda(CeedElemTopology, CeedInt, CeedInt,
+                                       CeedInt, const CeedScalar *,
+                                       const CeedScalar *, const CeedScalar *,
+                                       const CeedScalar *, CeedBasis);
 
-CEED_INTERN int CeedQFunctionCreate_Hip(CeedQFunction qf);
+CEED_INTERN int CeedQFunctionCreate_Cuda(CeedQFunction qf);
 
-CEED_INTERN int CeedQFunctionContextCreate_Hip(CeedQFunctionContext ctx);
+CEED_INTERN int CeedQFunctionContextCreate_Cuda(CeedQFunctionContext ctx);
 
-CEED_INTERN int CeedOperatorCreate_Hip(CeedOperator op);
+CEED_INTERN int CeedOperatorCreate_Cuda(CeedOperator op);
 
-CEED_INTERN int CeedCompositeOperatorCreate_Hip(CeedOperator op);
+CEED_INTERN int CeedCompositeOperatorCreate_Cuda(CeedOperator op);
 #endif
