+++ backends/cuda-ref/ceed-cuda-vector.c
@@ -20,7 +20,7 @@
 #include <cuda_runtime.h>
 #include <math.h>
 #include <string.h>
-#include "ceed-cuda.h"
+#include "ceed-cuda-ref.h"
 
 //------------------------------------------------------------------------------
 // * Bytes used
