+++ backends/cuda/ceed-cuda-common.c
@@ -23,19 +23,19 @@
 int CeedCudaInit(Ceed ceed, const char *resource) {
   int ierr;
   const char *device_spec = strstr(resource, ":device_id=");
-  const int deviceID = (device_spec) ? atoi(device_spec+11) : -1;
+  const int device_id = (device_spec) ? atoi(device_spec + 11) : -1;
 
-  int currentDeviceID;
-  ierr = cudaGetDevice(&currentDeviceID); CeedChk_Cu(ceed,ierr);
-  if (deviceID >= 0 && currentDeviceID != deviceID) {
-    ierr = cudaSetDevice(deviceID); CeedChk_Cu(ceed,ierr);
-    currentDeviceID = deviceID;
+  int current_device_id;
+  ierr = cudaGetDevice(&current_device_id); CeedChk_Cu(ceed, ierr);
+  if (device_id >= 0 && current_device_id != device_id) {
+    ierr = cudaSetDevice(device_id); CeedChk_Cu(ceed, ierr);
+    current_device_id = device_id;
   }
   Ceed_Cuda *data;
   ierr = CeedGetData(ceed, &data); CeedChkBackend(ierr);
-  data->deviceId = currentDeviceID;
-  ierr = cudaGetDeviceProperties(&data->deviceProp, currentDeviceID);
-  CeedChk_Cu(ceed,ierr);
+  data->device_id = current_device_id;
+  ierr = cudaGetDeviceProperties(&data->device_prop, current_device_id);
+  CeedChk_Cu(ceed, ierr);
   return CEED_ERROR_SUCCESS;
 }
 
@@ -46,8 +46,8 @@ int CeedDestroy_Cuda(Ceed ceed) {
   int ierr;
   Ceed_Cuda *data;
   ierr = CeedGetData(ceed, &data); CeedChkBackend(ierr);
-  if (data->cublasHandle) {
-    ierr = cublasDestroy(data->cublasHandle); CeedChk_Cublas(ceed, ierr);
+  if (data->cublas_handle) {
+    ierr = cublasDestroy(data->cublas_handle); CeedChk_Cublas(ceed, ierr);
   }
   ierr = CeedFree(&data); CeedChkBackend(ierr);
   return CEED_ERROR_SUCCESS;
