+++ backends/cuda/ceed-cuda-compile.h
@@ -21,26 +21,30 @@
 #include <cuda.h>
 #include <nvrtc.h>
 
+static inline CeedInt CeedDivUpInt(CeedInt numerator, CeedInt denominator) {
+  return (numerator + denominator - 1) / denominator;
+}
+
 CEED_INTERN int CeedCompileCuda(Ceed ceed, const char *source, CUmodule *module,
-                                const CeedInt numopts, ...);
+                                const CeedInt num_opts, ...);
 
 CEED_INTERN int CeedGetKernelCuda(Ceed ceed, CUmodule module, const char *name,
                                   CUfunction *kernel);
 
 CEED_INTERN int CeedRunKernelCuda(Ceed ceed, CUfunction kernel,
-                                  const int gridSize,
-                                  const int blockSize, void **args);
+                                  const int grid_size,
+                                  const int block_size, void **args);
 
 CEED_INTERN int CeedRunKernelAutoblockCuda(Ceed ceed, CUfunction kernel,
     size_t size, void **args);
 
 CEED_INTERN int CeedRunKernelDimCuda(Ceed ceed, CUfunction kernel,
-                                     const int gridSize,
-                                     const int blockSizeX, const int blockSizeY,
-                                     const int blockSizeZ, void **args);
+                                     const int grid_size,
+                                     const int block_size_x, const int block_size_y,
+                                     const int block_size_z, void **args);
 
 CEED_INTERN int CeedRunKernelDimSharedCuda(Ceed ceed, CUfunction kernel,
-    const int gridSize, const int blockSizeX, const int blockSizeY,
-    const int blockSizeZ, const int sharedMemSize, void **args);
+    const int grid_size, const int block_size_x, const int block_size_y,
+    const int block_size_z, const int shared_mem_size, void **args);
 
 #endif // _ceed_cuda_compile_h
