+++ backends/hip-ref/ceed-hip-ref-basis.c
@@ -17,7 +17,7 @@
 #include <ceed/ceed.h>
 #include <ceed/backend.h>
 #include <hip/hip_runtime.h>
-#include "ceed-hip.h"
+#include "ceed-hip-ref.h"
 #include "../hip/ceed-hip-compile.h"
 
 //------------------------------------------------------------------------------
