+++ backends/hip-ref/ceed-hip-ref-operator.c
@@ -16,20 +16,19 @@
 
 #include <ceed/ceed.h>
 #include <ceed/backend.h>
+#include <hip/hip_runtime.h>
 #include <assert.h>
-#include <cuda.h>
-#include <cuda_runtime.h>
 #include <stdbool.h>
 #include <string.h>
-#include "ceed-cuda.h"
-#include "../cuda/ceed-cuda-compile.h"
+#include "ceed-hip-ref.h"
+#include "../hip/ceed-hip-compile.h"
 
 //------------------------------------------------------------------------------
 // Destroy operator
 //------------------------------------------------------------------------------
-static int CeedOperatorDestroy_Cuda(CeedOperator op) {
+static int CeedOperatorDestroy_Hip(CeedOperator op) {
   int ierr;
-  CeedOperator_Cuda *impl;
+  CeedOperator_Hip *impl;
   ierr = CeedOperatorGetData(op, &impl); CeedChkBackend(ierr);
 
   // Apply data
@@ -48,7 +47,7 @@ static int CeedOperatorDestroy_Cuda(CeedOperator op) {
   }
   ierr = CeedFree(&impl->qvecsout); CeedChkBackend(ierr);
 
-  // QFunction assembly data
+  // QFunction diagonal assembly data
   for (CeedInt i=0; i<impl->qfnumactivein; i++) {
     ierr = CeedVectorDestroy(&impl->qfactivein[i]); CeedChkBackend(ierr);
   }
@@ -58,16 +57,16 @@ static int CeedOperatorDestroy_Cuda(CeedOperator op) {
   if (impl->diag) {
     Ceed ceed;
     ierr = CeedOperatorGetCeed(op, &ceed); CeedChkBackend(ierr);
-    CeedChk_Cu(ceed, cuModuleUnload(impl->diag->module));
+    CeedChk_Hip(ceed, hipModuleUnload(impl->diag->module));
     ierr = CeedFree(&impl->diag->h_emodein); CeedChkBackend(ierr);
     ierr = CeedFree(&impl->diag->h_emodeout); CeedChkBackend(ierr);
-    ierr = cudaFree(impl->diag->d_emodein); CeedChk_Cu(ceed, ierr);
-    ierr = cudaFree(impl->diag->d_emodeout); CeedChk_Cu(ceed, ierr);
-    ierr = cudaFree(impl->diag->d_identity); CeedChk_Cu(ceed, ierr);
-    ierr = cudaFree(impl->diag->d_interpin); CeedChk_Cu(ceed, ierr);
-    ierr = cudaFree(impl->diag->d_interpout); CeedChk_Cu(ceed, ierr);
-    ierr = cudaFree(impl->diag->d_gradin); CeedChk_Cu(ceed, ierr);
-    ierr = cudaFree(impl->diag->d_gradout); CeedChk_Cu(ceed, ierr);
+    ierr = hipFree(impl->diag->d_emodein); CeedChk_Hip(ceed, ierr);
+    ierr = hipFree(impl->diag->d_emodeout); CeedChk_Hip(ceed, ierr);
+    ierr = hipFree(impl->diag->d_identity); CeedChk_Hip(ceed, ierr);
+    ierr = hipFree(impl->diag->d_interpin); CeedChk_Hip(ceed, ierr);
+    ierr = hipFree(impl->diag->d_interpout); CeedChk_Hip(ceed, ierr);
+    ierr = hipFree(impl->diag->d_gradin); CeedChk_Hip(ceed, ierr);
+    ierr = hipFree(impl->diag->d_gradout); CeedChk_Hip(ceed, ierr);
     ierr = CeedElemRestrictionDestroy(&impl->diag->pbdiagrstr);
     CeedChkBackend(ierr);
     ierr = CeedVectorDestroy(&impl->diag->elemdiag); CeedChkBackend(ierr);
@@ -82,11 +81,11 @@ static int CeedOperatorDestroy_Cuda(CeedOperator op) {
 //------------------------------------------------------------------------------
 // Setup infields or outfields
 //------------------------------------------------------------------------------
-static int CeedOperatorSetupFields_Cuda(CeedQFunction qf, CeedOperator op,
-                                        bool isinput, CeedVector *evecs,
-                                        CeedVector *qvecs, CeedInt starte,
-                                        CeedInt numfields, CeedInt Q,
-                                        CeedInt numelements) {
+static int CeedOperatorSetupFields_Hip(CeedQFunction qf, CeedOperator op,
+                                       bool isinput, CeedVector *evecs,
+                                       CeedVector *qvecs, CeedInt starte,
+                                       CeedInt numfields, CeedInt Q,
+                                       CeedInt numelements) {
   CeedInt dim, ierr, size;
   Ceed ceed;
   ierr = CeedOperatorGetCeed(op, &ceed); CeedChkBackend(ierr);
@@ -191,7 +190,7 @@ static int CeedOperatorSetupFields_Cuda(CeedQFunction qf, CeedOperator op,
 // CeedOperator needs to connect all the named fields (be they active or passive)
 //   to the named inputs and outputs of its CeedQFunction.
 //------------------------------------------------------------------------------
-static int CeedOperatorSetup_Cuda(CeedOperator op) {
+static int CeedOperatorSetup_Hip(CeedOperator op) {
   int ierr;
   bool setupdone;
   ierr = CeedOperatorIsSetupDone(op, &setupdone); CeedChkBackend(ierr);
@@ -199,14 +198,13 @@ static int CeedOperatorSetup_Cuda(CeedOperator op) {
     return CEED_ERROR_SUCCESS;
   Ceed ceed;
   ierr = CeedOperatorGetCeed(op, &ceed); CeedChkBackend(ierr);
-  CeedOperator_Cuda *impl;
+  CeedOperator_Hip *impl;
   ierr = CeedOperatorGetData(op, &impl); CeedChkBackend(ierr);
   CeedQFunction qf;
   ierr = CeedOperatorGetQFunction(op, &qf); CeedChkBackend(ierr);
   CeedInt Q, numelements, numinputfields, numoutputfields;
   ierr = CeedOperatorGetNumQuadraturePoints(op, &Q); CeedChkBackend(ierr);
   ierr = CeedOperatorGetNumElements(op, &numelements); CeedChkBackend(ierr);
-  CeedChkBackend(ierr);
   CeedOperatorField *opinputfields, *opoutputfields;
   ierr = CeedOperatorGetFields(op, &numinputfields, &opinputfields,
                                &numoutputfields, &opoutputfields);
@@ -226,16 +224,16 @@ static int CeedOperatorSetup_Cuda(CeedOperator op) {
 
   // Set up infield and outfield evecs and qvecs
   // Infields
-  ierr = CeedOperatorSetupFields_Cuda(qf, op, true,
-                                      impl->evecs, impl->qvecsin, 0,
-                                      numinputfields, Q, numelements);
+  ierr = CeedOperatorSetupFields_Hip(qf, op, true,
+                                     impl->evecs, impl->qvecsin, 0,
+                                     numinputfields, Q, numelements);
   CeedChkBackend(ierr);
 
   // Outfields
-  ierr = CeedOperatorSetupFields_Cuda(qf, op, false,
-                                      impl->evecs, impl->qvecsout,
-                                      numinputfields, numoutputfields, Q,
-                                      numelements); CeedChkBackend(ierr);
+  ierr = CeedOperatorSetupFields_Hip(qf, op, false,
+                                     impl->evecs, impl->qvecsout,
+                                     numinputfields, numoutputfields, Q,
+                                     numelements); CeedChkBackend(ierr);
 
   ierr = CeedOperatorSetSetupDone(op); CeedChkBackend(ierr);
   return CEED_ERROR_SUCCESS;
@@ -244,10 +242,10 @@ static int CeedOperatorSetup_Cuda(CeedOperator op) {
 //------------------------------------------------------------------------------
 // Setup Operator Inputs
 //------------------------------------------------------------------------------
-static inline int CeedOperatorSetupInputs_Cuda(CeedInt numinputfields,
+static inline int CeedOperatorSetupInputs_Hip(CeedInt numinputfields,
     CeedQFunctionField *qfinputfields, CeedOperatorField *opinputfields,
     CeedVector invec, const bool skipactive, CeedScalar *edata[2*CEED_FIELD_MAX],
-    CeedOperator_Cuda *impl, CeedRequest *request) {
+    CeedOperator_Hip *impl, CeedRequest *request) {
   CeedInt ierr;
   CeedEvalMode emode;
   CeedVector vec;
@@ -296,10 +294,10 @@ static inline int CeedOperatorSetupInputs_Cuda(CeedInt numinputfields,
 //------------------------------------------------------------------------------
 // Input Basis Action
 //------------------------------------------------------------------------------
-static inline int CeedOperatorInputBasis_Cuda(CeedInt numelements,
+static inline int CeedOperatorInputBasis_Hip(CeedInt numelements,
     CeedQFunctionField *qfinputfields, CeedOperatorField *opinputfields,
     CeedInt numinputfields, const bool skipactive,
-    CeedScalar *edata[2*CEED_FIELD_MAX],CeedOperator_Cuda *impl) {
+    CeedScalar *edata[2*CEED_FIELD_MAX], CeedOperator_Hip *impl) {
   CeedInt ierr;
   CeedInt elemsize, size;
   CeedElemRestriction Erestrict;
@@ -356,10 +354,10 @@ static inline int CeedOperatorInputBasis_Cuda(CeedInt numelements,
 //------------------------------------------------------------------------------
 // Restore Input Vectors
 //------------------------------------------------------------------------------
-static inline int CeedOperatorRestoreInputs_Cuda(CeedInt numinputfields,
+static inline int CeedOperatorRestoreInputs_Hip(CeedInt numinputfields,
     CeedQFunctionField *qfinputfields, CeedOperatorField *opinputfields,
     const bool skipactive, CeedScalar *edata[2*CEED_FIELD_MAX],
-    CeedOperator_Cuda *impl) {
+    CeedOperator_Hip *impl) {
   CeedInt ierr;
   CeedEvalMode emode;
   CeedVector vec;
@@ -393,17 +391,16 @@ static inline int CeedOperatorRestoreInputs_Cuda(CeedInt numinputfields,
 //------------------------------------------------------------------------------
 // Apply and add to output
 //------------------------------------------------------------------------------
-static int CeedOperatorApplyAdd_Cuda(CeedOperator op, CeedVector invec,
-                                     CeedVector outvec, CeedRequest *request) {
+static int CeedOperatorApplyAdd_Hip(CeedOperator op, CeedVector invec,
+                                    CeedVector outvec, CeedRequest *request) {
   int ierr;
-  CeedOperator_Cuda *impl;
+  CeedOperator_Hip *impl;
   ierr = CeedOperatorGetData(op, &impl); CeedChkBackend(ierr);
   CeedQFunction qf;
   ierr = CeedOperatorGetQFunction(op, &qf); CeedChkBackend(ierr);
   CeedInt Q, numelements, elemsize, numinputfields, numoutputfields, size;
   ierr = CeedOperatorGetNumQuadraturePoints(op, &Q); CeedChkBackend(ierr);
   ierr = CeedOperatorGetNumElements(op, &numelements); CeedChkBackend(ierr);
-  CeedChkBackend(ierr);
   CeedOperatorField *opinputfields, *opoutputfields;
   ierr = CeedOperatorGetFields(op, &numinputfields, &opinputfields,
                                &numoutputfields, &opoutputfields);
@@ -415,19 +412,19 @@ static int CeedOperatorApplyAdd_Cuda(CeedOperator op, CeedVector invec,
   CeedVector vec;
   CeedBasis basis;
   CeedElemRestriction Erestrict;
-  CeedScalar *edata[2*CEED_FIELD_MAX] = {0};
+  CeedScalar *edata[2*CEED_FIELD_MAX];
 
   // Setup
-  ierr = CeedOperatorSetup_Cuda(op); CeedChkBackend(ierr);
+  ierr = CeedOperatorSetup_Hip(op); CeedChkBackend(ierr);
 
   // Input Evecs and Restriction
-  ierr = CeedOperatorSetupInputs_Cuda(numinputfields, qfinputfields,
-                                      opinputfields, invec, false, edata,
-                                      impl, request); CeedChkBackend(ierr);
+  ierr = CeedOperatorSetupInputs_Hip(numinputfields, qfinputfields,
+                                     opinputfields, invec, false, edata,
+                                     impl, request); CeedChkBackend(ierr);
 
   // Input basis apply if needed
-  ierr = CeedOperatorInputBasis_Cuda(numelements, qfinputfields, opinputfields,
-                                     numinputfields, false, edata, impl);
+  ierr = CeedOperatorInputBasis_Hip(numelements, qfinputfields, opinputfields,
+                                    numinputfields, false, edata, impl);
   CeedChkBackend(ierr);
 
   // Output pointers, as necessary
@@ -519,8 +516,8 @@ static int CeedOperatorApplyAdd_Cuda(CeedOperator op, CeedVector invec,
   }
 
   // Restore input arrays
-  ierr = CeedOperatorRestoreInputs_Cuda(numinputfields, qfinputfields,
-                                        opinputfields, false, edata, impl);
+  ierr = CeedOperatorRestoreInputs_Hip(numinputfields, qfinputfields,
+                                       opinputfields, false, edata, impl);
   CeedChkBackend(ierr);
   return CEED_ERROR_SUCCESS;
 }
@@ -528,18 +525,17 @@ static int CeedOperatorApplyAdd_Cuda(CeedOperator op, CeedVector invec,
 //------------------------------------------------------------------------------
 // Core code for assembling linear QFunction
 //------------------------------------------------------------------------------
-static inline int CeedOperatorLinearAssembleQFunctionCore_Cuda(CeedOperator op,
+static inline int CeedOperatorLinearAssembleQFunctionCore_Hip(CeedOperator op,
     bool build_objects, CeedVector *assembled, CeedElemRestriction *rstr,
     CeedRequest *request) {
   int ierr;
-  CeedOperator_Cuda *impl;
+  CeedOperator_Hip *impl;
   ierr = CeedOperatorGetData(op, &impl); CeedChkBackend(ierr);
   CeedQFunction qf;
   ierr = CeedOperatorGetQFunction(op, &qf); CeedChkBackend(ierr);
   CeedInt Q, numelements, numinputfields, numoutputfields, size;
   ierr = CeedOperatorGetNumQuadraturePoints(op, &Q); CeedChkBackend(ierr);
   ierr = CeedOperatorGetNumElements(op, &numelements); CeedChkBackend(ierr);
-  CeedChkBackend(ierr);
   CeedOperatorField *opinputfields, *opoutputfields;
   ierr = CeedOperatorGetFields(op, &numinputfields, &opinputfields,
                                &numoutputfields, &opoutputfields);
@@ -559,7 +555,7 @@ static inline int CeedOperatorLinearAssembleQFunctionCore_Cuda(CeedOperator op,
   CeedScalar *edata[2*CEED_FIELD_MAX];
 
   // Setup
-  ierr = CeedOperatorSetup_Cuda(op); CeedChkBackend(ierr);
+  ierr = CeedOperatorSetup_Hip(op); CeedChkBackend(ierr);
 
   // Check for identity
   bool identityqf;
@@ -571,9 +567,9 @@ static inline int CeedOperatorLinearAssembleQFunctionCore_Cuda(CeedOperator op,
   // LCOV_EXCL_STOP
 
   // Input Evecs and Restriction
-  ierr = CeedOperatorSetupInputs_Cuda(numinputfields, qfinputfields,
-                                      opinputfields, NULL, true, edata,
-                                      impl, request); CeedChkBackend(ierr);
+  ierr = CeedOperatorSetupInputs_Hip(numinputfields, qfinputfields,
+                                     opinputfields, NULL, true, edata,
+                                     impl, request); CeedChkBackend(ierr);
 
   // Count number of active input fields
   if (!numactivein) {
@@ -643,8 +639,8 @@ static inline int CeedOperatorLinearAssembleQFunctionCore_Cuda(CeedOperator op,
   CeedChkBackend(ierr);
 
   // Input basis apply
-  ierr = CeedOperatorInputBasis_Cuda(numelements, qfinputfields, opinputfields,
-                                     numinputfields, true, edata, impl);
+  ierr = CeedOperatorInputBasis_Hip(numelements, qfinputfields, opinputfields,
+                                    numinputfields, true, edata, impl);
   CeedChkBackend(ierr);
 
   // Assemble QFunction
@@ -687,8 +683,8 @@ static inline int CeedOperatorLinearAssembleQFunctionCore_Cuda(CeedOperator op,
   }
 
   // Restore input arrays
-  ierr = CeedOperatorRestoreInputs_Cuda(numinputfields, qfinputfields,
-                                        opinputfields, true, edata, impl);
+  ierr = CeedOperatorRestoreInputs_Hip(numinputfields, qfinputfields,
+                                       opinputfields, true, edata, impl);
   CeedChkBackend(ierr);
 
   // Restore output
@@ -700,19 +696,19 @@ static inline int CeedOperatorLinearAssembleQFunctionCore_Cuda(CeedOperator op,
 //------------------------------------------------------------------------------
 // Assemble Linear QFunction
 //------------------------------------------------------------------------------
-static int CeedOperatorLinearAssembleQFunction_Cuda(CeedOperator op,
+static int CeedOperatorLinearAssembleQFunction_Hip(CeedOperator op,
     CeedVector *assembled, CeedElemRestriction *rstr, CeedRequest *request) {
-  return CeedOperatorLinearAssembleQFunctionCore_Cuda(op, true, assembled, rstr,
+  return CeedOperatorLinearAssembleQFunctionCore_Hip(op, true, assembled, rstr,
          request);
 }
 
 //------------------------------------------------------------------------------
-// Update Assembled Linear QFunction
+// Assemble Linear QFunction
 //------------------------------------------------------------------------------
-static int CeedOperatorLinearAssembleQFunctionUpdate_Cuda(CeedOperator op,
+static int CeedOperatorLinearAssembleQFunctionUpdate_Hip(CeedOperator op,
     CeedVector assembled, CeedElemRestriction rstr, CeedRequest *request) {
-  return CeedOperatorLinearAssembleQFunctionCore_Cuda(op, false, &assembled,
-         &rstr, request);
+  return CeedOperatorLinearAssembleQFunctionCore_Hip(op, false, &assembled, &rstr,
+         request);
 }
 
 //------------------------------------------------------------------------------
@@ -740,7 +736,7 @@ typedef enum {
 //------------------------------------------------------------------------------
 // Get Basis Emode Pointer
 //------------------------------------------------------------------------------
-extern "C" __device__ void CeedOperatorGetBasisPointer_Cuda(const CeedScalar **basisptr,
+extern "C" __device__ void CeedOperatorGetBasisPointer_Hip(const CeedScalar **basisptr,
     CeedEvalMode emode, const CeedScalar *identity, const CeedScalar *interp,
     const CeedScalar *grad) {
   switch (emode) {
@@ -784,14 +780,14 @@ __device__ void diagonalCore(const CeedInt nelem,
       const CeedScalar *bt = NULL;
       if (emodeout[eout] == CEED_EVAL_GRAD)
         dout += 1;
-      CeedOperatorGetBasisPointer_Cuda(&bt, emodeout[eout], identity, interpout,
+      CeedOperatorGetBasisPointer_Hip(&bt, emodeout[eout], identity, interpout,
                                       &gradout[dout*NQPTS*NNODES]);
       CeedInt din = -1;
       for (CeedInt ein = 0; ein < NUMEMODEIN; ein++) {
         const CeedScalar *b = NULL;
         if (emodein[ein] == CEED_EVAL_GRAD)
           din += 1;
-        CeedOperatorGetBasisPointer_Cuda(&b, emodein[ein], identity, interpin,
+        CeedOperatorGetBasisPointer_Hip(&b, emodein[ein], identity, interpin,
                                         &gradin[din*NQPTS*NNODES]);
         // Each component
         for (CeedInt compOut = 0; compOut < NCOMP; compOut++) {
@@ -902,7 +898,7 @@ static int CreatePBRestriction(CeedElemRestriction rstr,
 //------------------------------------------------------------------------------
 // Assemble diagonal setup
 //------------------------------------------------------------------------------
-static inline int CeedOperatorAssembleDiagonalSetup_Cuda(CeedOperator op,
+static inline int CeedOperatorAssembleDiagonalSetup_Hip(CeedOperator op,
     const bool pointBlock) {
   int ierr;
   Ceed ceed;
@@ -1011,10 +1007,10 @@ static inline int CeedOperatorAssembleDiagonalSetup_Cuda(CeedOperator op,
   }
 
   // Operator data struct
-  CeedOperator_Cuda *impl;
+  CeedOperator_Hip *impl;
   ierr = CeedOperatorGetData(op, &impl); CeedChkBackend(ierr);
   ierr = CeedCalloc(1, &impl->diag); CeedChkBackend(ierr);
-  CeedOperatorDiag_Cuda *diag = impl->diag;
+  CeedOperatorDiag_Hip *diag = impl->diag;
   diag->basisin = basisin;
   diag->basisout = basisout;
   diag->h_emodein = emodein;
@@ -1027,18 +1023,18 @@ static inline int CeedOperatorAssembleDiagonalSetup_Cuda(CeedOperator op,
   ierr = CeedBasisGetNumNodes(basisin, &nnodes); CeedChkBackend(ierr);
   ierr = CeedBasisGetNumQuadraturePoints(basisin, &nqpts); CeedChkBackend(ierr);
   diag->nnodes = nnodes;
-  ierr = CeedCompileCuda(ceed, diagonalkernels, &diag->module, 5,
-                         "NUMEMODEIN", numemodein,
-                         "NUMEMODEOUT", numemodeout,
-                         "NNODES", nnodes,
-                         "NQPTS", nqpts,
-                         "NCOMP", ncomp
-                        ); CeedChk_Cu(ceed, ierr);
-  ierr = CeedGetKernelCuda(ceed, diag->module, "linearDiagonal",
-                           &diag->linearDiagonal); CeedChk_Cu(ceed, ierr);
-  ierr = CeedGetKernelCuda(ceed, diag->module, "linearPointBlockDiagonal",
-                           &diag->linearPointBlock);
-  CeedChk_Cu(ceed, ierr);
+  ierr = CeedCompileHip(ceed, diagonalkernels, &diag->module, 5,
+                        "NUMEMODEIN", numemodein,
+                        "NUMEMODEOUT", numemodeout,
+                        "NNODES", nnodes,
+                        "NQPTS", nqpts,
+                        "NCOMP", ncomp
+                       ); CeedChk_Hip(ceed, ierr);
+  ierr = CeedGetKernelHip(ceed, diag->module, "linearDiagonal",
+                          &diag->linearDiagonal); CeedChk_Hip(ceed, ierr);
+  ierr = CeedGetKernelHip(ceed, diag->module, "linearPointBlockDiagonal",
+                          &diag->linearPointBlock);
+  CeedChk_Hip(ceed, ierr);
 
   // Basis matrices
   const CeedInt qBytes = nqpts * sizeof(CeedScalar);
@@ -1058,40 +1054,40 @@ static inline int CeedOperatorAssembleDiagonalSetup_Cuda(CeedOperator op,
     ierr = CeedCalloc(nqpts*nnodes, &identity); CeedChkBackend(ierr);
     for (CeedInt i=0; i<(nnodes<nqpts?nnodes:nqpts); i++)
       identity[i*nnodes+i] = 1.0;
-    ierr = cudaMalloc((void **)&diag->d_identity, iBytes); CeedChk_Cu(ceed, ierr);
-    ierr = cudaMemcpy(diag->d_identity, identity, iBytes,
-                      cudaMemcpyHostToDevice); CeedChk_Cu(ceed, ierr);
+    ierr = hipMalloc((void **)&diag->d_identity, iBytes); CeedChk_Hip(ceed, ierr);
+    ierr = hipMemcpy(diag->d_identity, identity, iBytes,
+                     hipMemcpyHostToDevice); CeedChk_Hip(ceed, ierr);
   }
 
   // CEED_EVAL_INTERP
   ierr = CeedBasisGetInterp(basisin, &interpin); CeedChkBackend(ierr);
-  ierr = cudaMalloc((void **)&diag->d_interpin, iBytes); CeedChk_Cu(ceed, ierr);
-  ierr = cudaMemcpy(diag->d_interpin, interpin, iBytes,
-                    cudaMemcpyHostToDevice); CeedChk_Cu(ceed, ierr);
+  ierr = hipMalloc((void **)&diag->d_interpin, iBytes); CeedChk_Hip(ceed, ierr);
+  ierr = hipMemcpy(diag->d_interpin, interpin, iBytes,
+                   hipMemcpyHostToDevice); CeedChk_Hip(ceed, ierr);
   ierr = CeedBasisGetInterp(basisout, &interpout); CeedChkBackend(ierr);
-  ierr = cudaMalloc((void **)&diag->d_interpout, iBytes); CeedChk_Cu(ceed, ierr);
-  ierr = cudaMemcpy(diag->d_interpout, interpout, iBytes,
-                    cudaMemcpyHostToDevice); CeedChk_Cu(ceed, ierr);
+  ierr = hipMalloc((void **)&diag->d_interpout, iBytes); CeedChk_Hip(ceed, ierr);
+  ierr = hipMemcpy(diag->d_interpout, interpout, iBytes,
+                   hipMemcpyHostToDevice); CeedChk_Hip(ceed, ierr);
 
   // CEED_EVAL_GRAD
   ierr = CeedBasisGetGrad(basisin, &gradin); CeedChkBackend(ierr);
-  ierr = cudaMalloc((void **)&diag->d_gradin, gBytes); CeedChk_Cu(ceed, ierr);
-  ierr = cudaMemcpy(diag->d_gradin, gradin, gBytes,
-                    cudaMemcpyHostToDevice); CeedChk_Cu(ceed, ierr);
+  ierr = hipMalloc((void **)&diag->d_gradin, gBytes); CeedChk_Hip(ceed, ierr);
+  ierr = hipMemcpy(diag->d_gradin, gradin, gBytes,
+                   hipMemcpyHostToDevice); CeedChk_Hip(ceed, ierr);
   ierr = CeedBasisGetGrad(basisout, &gradout); CeedChkBackend(ierr);
-  ierr = cudaMalloc((void **)&diag->d_gradout, gBytes); CeedChk_Cu(ceed, ierr);
-  ierr = cudaMemcpy(diag->d_gradout, gradout, gBytes,
-                    cudaMemcpyHostToDevice); CeedChk_Cu(ceed, ierr);
+  ierr = hipMalloc((void **)&diag->d_gradout, gBytes); CeedChk_Hip(ceed, ierr);
+  ierr = hipMemcpy(diag->d_gradout, gradout, gBytes,
+                   hipMemcpyHostToDevice); CeedChk_Hip(ceed, ierr);
 
   // Arrays of emodes
-  ierr = cudaMalloc((void **)&diag->d_emodein, numemodein * eBytes);
-  CeedChk_Cu(ceed, ierr);
-  ierr = cudaMemcpy(diag->d_emodein, emodein, numemodein * eBytes,
-                    cudaMemcpyHostToDevice); CeedChk_Cu(ceed, ierr);
-  ierr = cudaMalloc((void **)&diag->d_emodeout, numemodeout * eBytes);
-  CeedChk_Cu(ceed, ierr);
-  ierr = cudaMemcpy(diag->d_emodeout, emodeout, numemodeout * eBytes,
-                    cudaMemcpyHostToDevice); CeedChk_Cu(ceed, ierr);
+  ierr = hipMalloc((void **)&diag->d_emodein, numemodein * eBytes);
+  CeedChk_Hip(ceed, ierr);
+  ierr = hipMemcpy(diag->d_emodein, emodein, numemodein * eBytes,
+                   hipMemcpyHostToDevice); CeedChk_Hip(ceed, ierr);
+  ierr = hipMalloc((void **)&diag->d_emodeout, numemodeout * eBytes);
+  CeedChk_Hip(ceed, ierr);
+  ierr = hipMemcpy(diag->d_emodeout, emodeout, numemodeout * eBytes,
+                   hipMemcpyHostToDevice); CeedChk_Hip(ceed, ierr);
 
   // Restriction
   diag->diagrstr = rstrout;
@@ -1102,12 +1098,12 @@ static inline int CeedOperatorAssembleDiagonalSetup_Cuda(CeedOperator op,
 //------------------------------------------------------------------------------
 // Assemble diagonal common code
 //------------------------------------------------------------------------------
-static inline int CeedOperatorAssembleDiagonalCore_Cuda(CeedOperator op,
+static inline int CeedOperatorAssembleDiagonalCore_Hip(CeedOperator op,
     CeedVector assembled, CeedRequest *request, const bool pointBlock) {
   int ierr;
   Ceed ceed;
   ierr = CeedOperatorGetCeed(op, &ceed); CeedChkBackend(ierr);
-  CeedOperator_Cuda *impl;
+  CeedOperator_Hip *impl;
   ierr = CeedOperatorGetData(op, &impl); CeedChkBackend(ierr);
 
   // Assemble QFunction
@@ -1122,10 +1118,10 @@ static inline int CeedOperatorAssembleDiagonalCore_Cuda(CeedOperator op,
 
   // Setup
   if (!impl->diag) {
-    ierr = CeedOperatorAssembleDiagonalSetup_Cuda(op, pointBlock);
+    ierr = CeedOperatorAssembleDiagonalSetup_Hip(op, pointBlock);
     CeedChkBackend(ierr);
   }
-  CeedOperatorDiag_Cuda *diag = impl->diag;
+  CeedOperatorDiag_Hip *diag = impl->diag;
   assert(diag != NULL);
 
   // Restriction
@@ -1139,6 +1135,7 @@ static inline int CeedOperatorAssembleDiagonalCore_Cuda(CeedOperator op,
   // Create diagonal vector
   CeedVector elemdiag = pointBlock ? diag->pbelemdiag : diag->elemdiag;
   if (!elemdiag) {
+    // Element diagonal vector
     ierr = CeedElemRestrictionCreateVector(diagrstr, NULL, &elemdiag);
     CeedChkBackend(ierr);
     if (pointBlock)
@@ -1168,12 +1165,12 @@ static inline int CeedOperatorAssembleDiagonalCore_Cuda(CeedOperator op,
                   &assembledqfarray, &elemdiagarray
                  };
   if (pointBlock) {
-    ierr = CeedRunKernelDimCuda(ceed, diag->linearPointBlock, grid,
-                                diag->nnodes, 1, elemsPerBlock, args);
+    ierr = CeedRunKernelDimHip(ceed, diag->linearPointBlock, grid,
+                               diag->nnodes, 1, elemsPerBlock, args);
     CeedChkBackend(ierr);
   } else {
-    ierr = CeedRunKernelDimCuda(ceed, diag->linearDiagonal, grid,
-                                diag->nnodes, 1, elemsPerBlock, args);
+    ierr = CeedRunKernelDimHip(ceed, diag->linearDiagonal, grid,
+                               diag->nnodes, 1, elemsPerBlock, args);
     CeedChkBackend(ierr);
   }
 
@@ -1195,7 +1192,7 @@ static inline int CeedOperatorAssembleDiagonalCore_Cuda(CeedOperator op,
 //------------------------------------------------------------------------------
 // Assemble composite diagonal common code
 //------------------------------------------------------------------------------
-static inline int CeedOperatorLinearAssembleAddDiagonalCompositeCore_Cuda(
+static inline int CeedOperatorLinearAssembleAddDiagonalCompositeCore_Hip(
   CeedOperator op, CeedVector assembled, CeedRequest *request,
   const bool pointBlock) {
   int ierr;
@@ -1204,7 +1201,7 @@ static inline int CeedOperatorLinearAssembleAddDiagonalCompositeCore_Cuda(
   ierr = CeedOperatorGetNumSub(op, &numSub); CeedChkBackend(ierr);
   ierr = CeedOperatorGetSubList(op, &subOperators); CeedChkBackend(ierr);
   for (CeedInt i = 0; i < numSub; i++) {
-    ierr = CeedOperatorAssembleDiagonalCore_Cuda(subOperators[i], assembled,
+    ierr = CeedOperatorAssembleDiagonalCore_Hip(subOperators[i], assembled,
            request, pointBlock); CeedChkBackend(ierr);
   }
   return CEED_ERROR_SUCCESS;
@@ -1213,82 +1210,83 @@ static inline int CeedOperatorLinearAssembleAddDiagonalCompositeCore_Cuda(
 //------------------------------------------------------------------------------
 // Assemble Linear Diagonal
 //------------------------------------------------------------------------------
-static int CeedOperatorLinearAssembleAddDiagonal_Cuda(CeedOperator op,
+static int CeedOperatorLinearAssembleAddDiagonal_Hip(CeedOperator op,
     CeedVector assembled, CeedRequest *request) {
   int ierr;
   bool isComposite;
   ierr = CeedOperatorIsComposite(op, &isComposite); CeedChkBackend(ierr);
   if (isComposite) {
-    return CeedOperatorLinearAssembleAddDiagonalCompositeCore_Cuda(op, assembled,
+    return CeedOperatorLinearAssembleAddDiagonalCompositeCore_Hip(op, assembled,
            request, false);
   } else {
-    return CeedOperatorAssembleDiagonalCore_Cuda(op, assembled, request, false);
+    return CeedOperatorAssembleDiagonalCore_Hip(op, assembled, request, false);
   }
 }
 
 //------------------------------------------------------------------------------
 // Assemble Linear Point Block Diagonal
 //------------------------------------------------------------------------------
-static int CeedOperatorLinearAssembleAddPointBlockDiagonal_Cuda(CeedOperator op,
+static int CeedOperatorLinearAssembleAddPointBlockDiagonal_Hip(CeedOperator op,
     CeedVector assembled, CeedRequest *request) {
   int ierr;
   bool isComposite;
   ierr = CeedOperatorIsComposite(op, &isComposite); CeedChkBackend(ierr);
   if (isComposite) {
-    return CeedOperatorLinearAssembleAddDiagonalCompositeCore_Cuda(op, assembled,
+    return CeedOperatorLinearAssembleAddDiagonalCompositeCore_Hip(op, assembled,
            request, true);
   } else {
-    return CeedOperatorAssembleDiagonalCore_Cuda(op, assembled, request, true);
+    return CeedOperatorAssembleDiagonalCore_Hip(op, assembled, request, true);
   }
 }
+
 
 //------------------------------------------------------------------------------
 // Create operator
 //------------------------------------------------------------------------------
-int CeedOperatorCreate_Cuda(CeedOperator op) {
+int CeedOperatorCreate_Hip(CeedOperator op) {
   int ierr;
   Ceed ceed;
   ierr = CeedOperatorGetCeed(op, &ceed); CeedChkBackend(ierr);
-  CeedOperator_Cuda *impl;
+  CeedOperator_Hip *impl;
 
   ierr = CeedCalloc(1, &impl); CeedChkBackend(ierr);
   ierr = CeedOperatorSetData(op, impl); CeedChkBackend(ierr);
 
   ierr = CeedSetBackendFunction(ceed, "Operator", op, "LinearAssembleQFunction",
-                                CeedOperatorLinearAssembleQFunction_Cuda);
+                                CeedOperatorLinearAssembleQFunction_Hip);
   CeedChkBackend(ierr);
   ierr = CeedSetBackendFunction(ceed, "Operator", op,
                                 "LinearAssembleQFunctionUpdate",
-                                CeedOperatorLinearAssembleQFunctionUpdate_Cuda);
+                                CeedOperatorLinearAssembleQFunctionUpdate_Hip);
   CeedChkBackend(ierr);
   ierr = CeedSetBackendFunction(ceed, "Operator", op, "LinearAssembleAddDiagonal",
-                                CeedOperatorLinearAssembleAddDiagonal_Cuda);
+                                CeedOperatorLinearAssembleAddDiagonal_Hip);
   CeedChkBackend(ierr);
   ierr = CeedSetBackendFunction(ceed, "Operator", op,
                                 "LinearAssembleAddPointBlockDiagonal",
-                                CeedOperatorLinearAssembleAddPointBlockDiagonal_Cuda);
+                                CeedOperatorLinearAssembleAddPointBlockDiagonal_Hip);
   CeedChkBackend(ierr);
   ierr = CeedSetBackendFunction(ceed, "Operator", op, "ApplyAdd",
-                                CeedOperatorApplyAdd_Cuda); CeedChkBackend(ierr);
+                                CeedOperatorApplyAdd_Hip); CeedChkBackend(ierr);
   ierr = CeedSetBackendFunction(ceed, "Operator", op, "Destroy",
-                                CeedOperatorDestroy_Cuda); CeedChkBackend(ierr);
+                                CeedOperatorDestroy_Hip); CeedChkBackend(ierr);
   return CEED_ERROR_SUCCESS;
 }
 
 //------------------------------------------------------------------------------
 // Composite Operator Create
 //------------------------------------------------------------------------------
-int CeedCompositeOperatorCreate_Cuda(CeedOperator op) {
+int CeedCompositeOperatorCreate_Hip(CeedOperator op) {
   int ierr;
   Ceed ceed;
   ierr = CeedOperatorGetCeed(op, &ceed); CeedChkBackend(ierr);
 
   ierr = CeedSetBackendFunction(ceed, "Operator", op, "LinearAssembleAddDiagonal",
-                                CeedOperatorLinearAssembleAddDiagonal_Cuda);
+                                CeedOperatorLinearAssembleAddDiagonal_Hip);
   CeedChkBackend(ierr);
   ierr = CeedSetBackendFunction(ceed, "Operator", op,
                                 "LinearAssembleAddPointBlockDiagonal",
-                                CeedOperatorLinearAssembleAddPointBlockDiagonal_Cuda);
+                                CeedOperatorLinearAssembleAddPointBlockDiagonal_Hip);
   CeedChkBackend(ierr);
   return CEED_ERROR_SUCCESS;
 }
