+++ backends/hip/ceed-hip-common.c
@@ -26,23 +26,23 @@
 int CeedHipInit(Ceed ceed, const char *resource) {
   int ierr;
   const char *device_spec = strstr(resource, ":device_id=");
-  const int deviceID = (device_spec) ? atoi(device_spec+11) : -1;
+  const int device_id = (device_spec) ? atoi(device_spec + 11) : -1;
 
-  int currentDeviceID;
-  ierr = hipGetDevice(&currentDeviceID); CeedChk_Hip(ceed,ierr);
-  if (deviceID >= 0 && currentDeviceID != deviceID) {
-    ierr = hipSetDevice(deviceID); CeedChk_Hip(ceed,ierr);
-    currentDeviceID = deviceID;
+  int current_device_id;
+  ierr = hipGetDevice(&current_device_id); CeedChk_Hip(ceed, ierr);
+  if (device_id >= 0 && current_device_id != device_id) {
+    ierr = hipSetDevice(device_id); CeedChk_Hip(ceed, ierr);
+    current_device_id = device_id;
   }
 
-  struct hipDeviceProp_t deviceProp;
-  ierr = hipGetDeviceProperties(&deviceProp, currentDeviceID);
-  CeedChk_Hip(ceed,ierr);
+  struct hipDeviceProp_t device_prop;
+  ierr = hipGetDeviceProperties(&device_prop, current_device_id);
+  CeedChk_Hip(ceed, ierr);
 
   Ceed_Hip *data;
   ierr = CeedGetData(ceed, &data); CeedChkBackend(ierr);
-  data->deviceId = currentDeviceID;
-  data->optblocksize = 256;
+  data->device_id = current_device_id;
+  data->opt_block_size = 256;
   return CEED_ERROR_SUCCESS;
 }
 
@@ -53,8 +53,8 @@ int CeedDestroy_Hip(Ceed ceed) {
   int ierr;
   Ceed_Hip *data;
   ierr = CeedGetData(ceed, &data); CeedChkBackend(ierr);
-  if (data->hipblasHandle) {
-    ierr = hipblasDestroy(data->hipblasHandle); CeedChk_Hipblas(ceed, ierr);
+  if (data->hipblas_handle) {
+    ierr = hipblasDestroy(data->hipblas_handle); CeedChk_Hipblas(ceed, ierr);
   }
   ierr = CeedFree(&data); CeedChkBackend(ierr);
   return CEED_ERROR_SUCCESS;
