+++ backends/hip/ceed-hip-compile.h
@@ -20,26 +20,27 @@
 #include <ceed/ceed.h>
 #include <hip/hip_runtime.h>
 
+static inline CeedInt CeedDivUpInt(CeedInt numerator, CeedInt denominator) {
+  return (numerator + denominator - 1) / denominator;
+}
+
 CEED_INTERN int CeedCompileHip(Ceed ceed, const char *source,
-                               hipModule_t *module,
-                               const CeedInt numopts, ...);
+                               hipModule_t *module, const CeedInt numopts, ...);
 
 CEED_INTERN int CeedGetKernelHip(Ceed ceed, hipModule_t module,
-                                 const char *name,
-                                 hipFunction_t *kernel);
+                                 const char *name, hipFunction_t *kernel);
 
 CEED_INTERN int CeedRunKernelHip(Ceed ceed, hipFunction_t kernel,
-                                 const int gridSize,
-                                 const int blockSize, void **args);
+                                 const int grid_size,
+                                 const int block_size, void **args);
 
 CEED_INTERN int CeedRunKernelDimHip(Ceed ceed, hipFunction_t kernel,
-                                    const int gridSize,
-                                    const int blockSizeX, const int blockSizeY,
-                                    const int blockSizeZ, void **args);
+                                    const int grid_size,
+                                    const int block_size_x, const int block_size_y,
+                                    const int block_size_z, void **args);
 
 CEED_INTERN int CeedRunKernelDimSharedHip(Ceed ceed, hipFunction_t kernel,
-    const int gridSize, const int blockSizeX,
-    const int blockSizeY, const int blockSizeZ,
-    const int sharedMemSize, void **args);
+    const int grid_size, const int block_size_x, const int block_size_y,
+    const int block_size_z, const int shared_mem_size, void **args);
 
 #endif // _ceed_hip_compile_h
